refactor(job-client): Make job client easy to mock
