+++ iothub/service/tests/JobClient/DeviceJobParametersTest.cs
@@ -1,7 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace Microsoft.Azure.Devices.Api.Test.JobClient
+namespace Microsoft.Azure.Devices.Api.Test
 {
     using System;
     using System.Collections.Generic;
